# backend/05_APIs_Externas/api_rest/openapi.yaml
openapi: 3.0.3
info:
  title: METGO 3D API REST
  description: API agrometeorológica MVP — Valle de Aconcagua (Chile)
  version: 1.0.0-mvp
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
  - url: https://metgo-api.onrender.com
    description: Producción Render
...
paths:
...
  /api/agricola/{estacion_id}/{cultivo}/cronograma:
    get:
      tags: [agricola]
      summary: Cronograma de riego dinámico 7 días
      description: |
        Próximos 7 días de recomendación de riego por cultivo,
        ajustando por pronóstico de lluvia y riesgo de helada (OpenMeteo).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema:
            type: string
            enum: [quillota, los_nogales, hijuelas, limache, olmue]
        - name: cultivo
          in: path
          required: true
          schema:
            type: string
            enum: [palto, citricos, vid, tomate, lechuga, hortalizas, cereales]
      responses:
        "200":
          description: Cronograma generado
        "400":
          description: Cultivo no válido
...
  /api/datos/etl/sync:
    post:
      tags: [datos]
      summary: Sincronizar OpenMeteo + CSV hacia SQLite histórico
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dias: { type: integer, example: 14 }
                incluir_csv: { type: boolean, default: true }
                incluir_archive:
                  type: boolean
                  default: false
                  description: Si true, descarga OpenMeteo Archive (ERA5) y persiste con fuente openmeteo_archive
                anios_archive:
                  type: integer
                  default: 5
                  description: Años hacia atrás para archive (solo si incluir_archive)
      responses:
        "200":
          description: |
            Detalle por estación (histórico, pronóstico, helada_sync, serie_helada_sync)
            y estadísticas store (incl. heladas_pronostico)
        "401":
          description: Sin token
  /api/cron/sync:
    get:
      tags:
        - Automatización ETL
      summary: Sincroniza datos meteorológicos (Requiere token)
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: incluir_archive
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Si true, descarga OpenMeteo Archive (ERA5) además del sync diario
        - name: anios_archive
          in: query
          required: false
          schema:
            type: integer
            default: 5
          description: Años hacia atrás (solo con incluir_archive=true)
      responses:
        '200':
          description: Sincronización exitosa
        '401':
          description: No autorizado
  /api/ensemble:
    get:
      tags:
        - Modelos Predictivos
      summary: Consenso multi-modelo OpenMeteo (ECMWF/GFS/ICON) con caché last-good
      responses:
        '200':
          description: Datos del ensemble procesados (lista diaria)
        '503':
          description: OpenMeteo no disponible y sin caché last-good
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    fecha:
                      type: string
                      format: date
                    precipitacion:
                      type: object
                      properties:
                        mediana:
                          type: number
                        best_match:
                          type: number
                        probabilidad:
                          type: number
                    temperatura:
                      type: object
                      properties:
                        min_mediana:
                          type: number
                        alerta_helada:
                          type: boolean
                        tipo_helada:
                          type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT